refactor: 변수명, 함수 위치 기타사항  수정
