+++ src/Template.ts
@@ -1,15 +1,24 @@
-import addButtonImg from "../../templates/add-button.png";
-import categoryKorean from "../../templates/category-korean.png";
-import categoryAsian from "../../templates/category-asian.png";
-import categoryChinese from "../../templates/category-chinese.png";
-import categoryEtc from "../../templates/category-etc.png";
-import categoryJapanese from "../../templates/category-japanese.png";
-import categoryWestern from "../../templates/category-western.png";
+import addButtonImg from "../templates/add-button.png";
+import categoryKorean from "../templates/category-korean.png";
+import categoryAsian from "../templates/category-asian.png";
+import categoryChinese from "../templates/category-chinese.png";
+import categoryEtc from "../templates/category-etc.png";
+import categoryJapanese from "../templates/category-japanese.png";
+import categoryWestern from "../templates/category-western.png";
+
+export type CategoryType =
+  | "한식"
+  | "중식"
+  | "일식"
+  | "양식"
+  | "아시안"
+  | "기타";
+export type TakeTimeType = 5 | 10 | 15 | 20 | 30;
 
 export interface RestaurantType {
-  category: "한식" | "중식" | "일식" | "양식" | "아시안" | "기타";
+  category: CategoryType;
   name: string;
-  takeTime: 5 | 10 | 15 | 20 | 30;
+  takeTime: TakeTimeType;
   description?: string;
   link?: string;
 }
@@ -28,16 +37,33 @@ export const Template = {
     return `
     <header class="gnb">
       <h1 class="gnb__title text-title">점심 뭐 먹지</h1>
-      ${this.modalButton()}
+      ${this.BottomSheetButton()}
     </header>`;
   },
 
-  modalButton() {
+  BottomSheetButton() {
     return `
       <button type="button" class="gnb__button" aria-label="음식점 추가">
         <img src=${addButtonImg} alt="음식점 추가">
       </button>`;
   },
+
+  filterButton: `
+  <select name="category" id="category-filter" class="restaurant-filter">
+    <option value="전체">전체</option>
+    <option value="한식">한식</option>
+    <option value="중식">중식</option>
+    <option value="일식">일식</option>
+    <option value="양식">양식</option>
+    <option value="아시안">아시안</option>
+    <option value="기타">기타</option>
+  </select>`,
+
+  sortByButton: `
+  <select name="sorting" id="sorting-filter" class="restaurant-filter">
+    <option value="name">이름순</option>
+    <option value="distance">거리순</option>
+  </select>`,
 
   restaurantList(restaurantList: RestaurantType[]): string {
     return `<ul class='restaurant-list'>
@@ -60,28 +86,12 @@ export const Template = {
       <div class="restaurant__info">
         <h3 class="restaurant__name text-subtitle">${name}</h3>
         <span class="restaurant__distance text-body">캠퍼스부터 ${takeTime}분 내</span>
-      </div>
-    </li>${descriptionTemplate ? descriptionTemplate : ""}`;
+        ${descriptionTemplate ? descriptionTemplate : ""}
+        </div>
+    </li>`;
   },
 
-  filterButton: `
-    <select name="category" id="category-filter" class="restaurant-filter">
-      <option value="전체">전체</option>
-      <option value="한식">한식</option>
-      <option value="중식">중식</option>
-      <option value="일식">일식</option>
-      <option value="양식">양식</option>
-      <option value="아시안">아시안</option>
-      <option value="기타">기타</option>
-    </select>`,
-
-  sortByButton: `
-    <select name="sorting" id="sorting-filter" class="restaurant-filter">
-      <option value="name">이름순</option>
-      <option value="distance">거리순</option>
-    </select>`,
-
-  modalHeader: `<h2 class="modal-title text-title">새로운 음식점</h2>`,
+  bottomSheetHeader: `<h2 class="BottomSheet-title text-title">새로운 음식점</h2>`,
 
   categoryInput: `
     <div class="form-item form-item--required">
@@ -106,8 +116,8 @@ export const Template = {
 
   takeTimeInput: `
     <div class="form-item form-item--required">
-      <label for="distance text-caption">거리(도보 이동 시간) </label>
-      <select name="distance" id="distance" required>
+      <label for="takeTime text-caption">거리(도보 이동 시간) </label>
+      <select name="takeTime" id="takeTime" required>
         <option value="">선택해 주세요</option>
         <option value="5">5분 내</option>
         <option value="10">10분 내</option>
@@ -133,7 +143,7 @@ export const Template = {
 
   addButton: `
     <div class="button-container">
-      <button type="button" class="button button--secondary text-caption">취소하기</button>
-      <button class="button button--primary text-caption">추가하기</button>
+      <button type="reset" class="button button--secondary text-caption">취소하기</button>
+      <button type="submit" class="button button--primary text-caption">추가하기</button>
     </div>`,
 };
